ARROW-10070 Move count_set_bits tests to buffer.rs
